+++ R/dilplots-ggplot.R
@@ -818,14 +818,14 @@ plot_curve_ggplot <- function(
 #' @examples
 #'
 #' # Data Creation
-#' dilution_percent <- c(
+#' concentration <- c(
 #'   10, 20, 25, 40, 50, 60,
 #'   75, 80, 100, 125, 150,
 #'   10, 25, 40, 50, 60,
 #'   75, 80, 100, 125, 150
 #' )
 #'
-#' dilution_batch_name <- c(
+#' curve_batch_name <- c(
 #'   "B1", "B1", "B1", "B1", "B1",
 #'   "B1", "B1", "B1", "B1", "B1", "B1",
 #'   "B2", "B2", "B2", "B2", "B2",
@@ -843,7 +843,7 @@ plot_curve_ggplot <- function(
 #'   "Sample_125b", "Sample_150b"
 #' )
 #'
-#' lipid1_area_saturated <- c(
+#' curve_1_saturation_regime <- c(
 #'   5748124, 16616414, 21702718, 36191617,
 #'   49324541, 55618266, 66947588, 74964771,
 #'   75438063, 91770737, 94692060,
@@ -852,21 +852,21 @@ plot_curve_ggplot <- function(
 #'   78044338, 86158414
 #' )
 #'
-#' lipid2_area_linear <- c(
+#' curve_2_good_linearty <- c(
 #'   31538, 53709, 69990, 101977, 146436, 180960,
 #'   232881, 283780, 298289, 344519, 430432,
 #'   25463, 63387, 90624, 131274, 138069,
 #'   205353, 202407, 260205, 292257, 367924
 #' )
 #'
-#' lipid3_area_lod <- c(
+#' curve_3_noise_regime <- c(
 #'   544, 397, 829, 1437, 1808, 2231,
 #'   3343, 2915, 5268, 8031, 11045,
 #'   500, 903, 1267, 2031, 2100,
 #'   3563, 4500, 5300, 8500, 10430
 #' )
 #'
-#' lipid4_area_nonlinear <- c(
+#' curve_4_poor_linearty <- c(
 #'   380519, 485372, 478770, 474467, 531640, 576301,
 #'   501068, 550201, 515110, 499543, 474745,
 #'   197417, 322846, 478398, 423174, 418577,
@@ -875,16 +875,16 @@ plot_curve_ggplot <- function(
 #'
 #' curve_batch_annot <- tibble::tibble(
 #'   Sample_Name = sample_name,
-#'   Dilution_Batch_Name = dilution_batch_name,
-#'   Dilution_Percent = dilution_percent
+#'   Curve_Batch_Name = curve_batch_name,
+#'   Concentration = concentration
 #' )
 #'
 #' curve_data <- tibble::tibble(
 #'   Sample_Name = sample_name,
-#'   Lipid1 = lipid1_area_saturated,
-#'   Lipid2 = lipid2_area_linear,
-#'   Lipid3 = lipid3_area_lod,
-#'   Lipid4 = lipid4_area_nonlinear
+#'   `Curve_1` = curve_1_saturation_regime,
+#'   `Curve_2` = curve_2_good_linearty,
+#'   `Curve_3` = curve_3_noise_regime,
+#'   `Curve_4` = curve_4_poor_linearty
 #' )
 #'
 #' # Create curve table
@@ -892,24 +892,24 @@ plot_curve_ggplot <- function(
 #'   curve_batch_annot = curve_batch_annot,
 #'   curve_data_wide = curve_data,
 #'   common_column = "Sample_Name",
-#'   signal_var = "Area",
-#'   column_group = "Transition_Name"
+#'   signal_var = "Signal",
+#'   column_group = "Curve_Name"
 #' )
 #'
 #' # Create curve statistical summary
 #' curve_summary <- curve_table %>%
 #'   summarise_curve_table(
 #'     grouping_variable = c(
-#'       "Transition_Name",
-#'       "Dilution_Batch_Name"
+#'       "Curve_Name",
+#'       "Curve_Batch_Name"
 #'     ),
-#'     conc_var = "Dilution_Percent",
-#'     signal_var = "Area"
+#'     conc_var = "Concentration",
+#'     signal_var = "Signal"
 #'   ) %>%
-#'   dplyr::arrange(.data$Transition_Name) %>%
+#'   dplyr::arrange(.data[["Curve_Name"]]) %>%
 #'   evaluate_linearity(grouping_variable = c(
-#'     "Transition_Name",
-#'     "Dilution_Batch_Name"
+#'     "Curve_Name",
+#'     "Curve_Batch_Name"
 #'   ))
 #'
 #' # Create a ggplot table
@@ -931,18 +931,18 @@ add_ggplot_panel <- function(
     curve_summary = NULL,
     dilution_table = lifecycle::deprecated(),
     dilution_summary = lifecycle::deprecated(),
-    grouping_variable = c("Transition_Name",
-                          "Dilution_Batch_Name"),
-    curve_batch_var = "Dilution_Batch_Name",
+    grouping_variable = c("Curve_Name",
+                          "Curve_Batch_Name"),
+    curve_batch_var = "Curve_Batch_Name",
     curve_batch_col = c("#377eb8", "#4daf4a",
                        "#9C27B0", "#BCAAA4",
                        "#FF8A65", "#EFBBCF"),
     dil_batch_var = lifecycle::deprecated(),
     dil_batch_col = lifecycle::deprecated(),
-    conc_var = "Dilution_Percent",
+    conc_var = "Concentration",
     conc_var_units = "%",
     conc_var_interval = 50,
-    signal_var = "Area",
+    signal_var = "Signal",
     have_plot_title = TRUE,
     plot_summary_table = TRUE,
     plot_first_half_lin_reg = FALSE,